ID3v2: Rename items/identifier.rs
