+++ src/id3/v2/items/mod.rs
@@ -4,7 +4,6 @@ mod encapsulated_object;
 mod event_timing_codes_frame;
 mod extended_text_frame;
 mod extended_url_frame;
-mod identifier;
 mod key_value_frame;
 pub(in crate::id3::v2) mod language_frame;
 mod ownership_frame;
@@ -13,6 +12,7 @@ mod private_frame;
 mod relative_volume_adjustment_frame;
 mod sync_text;
 mod text_information_frame;
+mod unique_file_identifier;
 mod url_link_frame;
 
 pub use attached_picture_frame::AttachedPictureFrame;
@@ -21,7 +21,6 @@ pub use encapsulated_object::GeneralEncapsulatedObject;
 pub use event_timing_codes_frame::{Event, EventTimingCodesFrame, EventType};
 pub use extended_text_frame::ExtendedTextFrame;
 pub use extended_url_frame::ExtendedUrlFrame;
-pub use identifier::UniqueFileIdentifierFrame;
 pub use key_value_frame::KeyValueFrame;
 pub use language_frame::{CommentFrame, UnsynchronizedTextFrame};
 pub use ownership_frame::OwnershipFrame;
@@ -32,4 +31,5 @@ pub use relative_volume_adjustment_frame::{
 };
 pub use sync_text::{SyncTextContentType, SynchronizedText, TimestampFormat};
 pub use text_information_frame::TextInformationFrame;
+pub use unique_file_identifier::UniqueFileIdentifierFrame;
 pub use url_link_frame::UrlLinkFrame;
